backend/pyproject.toml
[project]
name = "miroshark-backend"
version = "0.1.0"
description = "MiroShark - A Simple and Universal Swarm Intelligence Engine, Predicting Anything"
requires-python = ">=3.11"
license = { text = "AGPL-3.0" }
authors = [
    { name = "MiroShark Team" }
]

dependencies = [
    # Core framework
    "flask>=3.0.0",
    "flask-cors>=6.0.0",

    # LLM related
    "openai>=1.0.0",

    # Neo4j
    "neo4j>=5.15.0",

    # HTTP requests (for Ollama embedding API)
    "requests>=2.31.0",

    # OASIS social media simulation (Wonderwall fork)
    # Wonderwall (oasis package) is bundled directly in backend/oasis/
    # "camel-oasis==0.2.5",  # replaced by bundled Wonderwall
    "camel-ai==0.2.78",

    # Wonderwall runtime dependencies (used by recsys, agent_graph, agents_generator)
    # These are hard imports that crash on startup if missing.
    "numpy>=1.26.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "torch>=2.0.0",
    "sentence-transformers>=2.2.0",
    "igraph>=0.11.0",

    # File processing
    "PyMuPDF>=1.24.0",
    # Encoding detection (supports non-UTF-8 encoded text files)
    "charset-normalizer>=3.0.0",
    "chardet>=5.0.0",

    # Utility libraries
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pipreqs>=0.5.0",
]

[tool.uv]
override-dependencies = [
    "pillow>=12.0.0",   # camel-oasis pins 10.3.0 which doesn't build on Python 3.13
    "tiktoken>=0.8.0",  # camel-ai pins 0.7.0 which doesn't build on Python 3.13
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
